Split coalescing batch construction
